requirements.txt
certifi==2020.12.5
chardet==4.0.0
configparser==5.0.2
idna==2.10
imageio==2.9.0
numpy==1.20.1
Pillow==8.1.2
pygifsicle==1.0.2
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
six==1.15.0
twilio==6.53.0
urllib3==1.26.3
